refactor: extract module definitions to separate module.rs files
- Move ApiIngress module definition from lib.rs to module.rs
- Move GrpcHub module definition from lib.rs to module.rs
- Move ModuleOrchestrator module definition from lib.rs to module.rs
- Update ARCHITECTURE_MANIFEST.md to reflect consistent module.rs pattern
- Maintain lib.rs as public exports and module organization only
